09/11_Home

+++ Libft/libft.h
@@ -13,7 +13,8 @@
 #ifndef LIBFT_H
 #define LIBFT_H
 #include <unistd.h>
-#include <stdio.h>
+#include <stdlib.h>
+#include <stdio.h> // to be removed!
 
 int	ft_atoi(const char *str);
 int	ft_isalnum(int i);
@@ -21,7 +22,7 @@ int	ft_isalpha(int i);
 int	ft_isascii(int i);
 int	ft_isdigit(int i);
 int	ft_isprint(int i);
-int	ft_strlen(const char *str);
+size_t	ft_strlen(const char *str);
 int	ft_tolower(int i);
 int	ft_toupper(int i);
